gcp_ai_pdf_summary_manual/requirements.txt
functions-framework
google-cloud-storage
google-cloud-aiplatform
pdfplumber